wsd: move WopiProxy to wopi/
