+++ wsd/ClientRequestDispatcher.cpp
@@ -26,10 +26,10 @@
 #include <FileServer.hpp>
 #if !MOBILEAPP
 #include <HostUtil.hpp>
+#include <wopi/WopiProxy.hpp>
 #endif // !MOBILEAPP
 #include <RequestDetails.hpp>
 #include <ProxyRequestHandler.hpp>
-#include <WopiProxy.hpp>
 #include <net/HttpHelper.hpp>
 #include <ConfigUtil.hpp>
 #include <JailUtil.hpp>
